+++ cmake/share/AllLoadC.cmake
@@ -35,7 +35,7 @@ if( NOT __ALL_LOAD_C_CMAKE__)
    # to generate an RPATH
    #
    function( CreateForceAllLoadList listname outputname)
-      if( ${${listname}})
+      if( ${listname})
          set( list ${BEGIN_ALL_LOAD})
          foreach( library ${${listname}})
             if( APPLE)